+++ src/effects/specialMonsters.js
@@ -17,7 +17,7 @@ export function baseMonsterFeatureEffect(document, label) {
     flags: {
       dae: {
         transfer: false,
-        stackable: false,
+        stackable: "none",
       },
       ddbimporter: {
         disabled: false,