+++ CHANGELOG.md
@@ -1,9 +1,14 @@
+# 2.9.55
+
+* Effect: Uncanny Dodge was no longer working.
+* Effect: Support Blessed Healer feature.
+
 # 2.9.54
 
 * Some fixes and extra debugging for Update to DDB functionality.
 * Some updates did not work with certain midi-effects.
 * If creating Midi-QOL effects on monsters, will now generate Legendary Resistance effect.
-* ECharacter Effect: Indomitable
+* Character Effect: Indomitable
 
 # 2.9.53